+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### v2.0.0
+
+- [x] Refactor Tracker class to use [x-rax](https://www.npmjs.com/package/x-ray)
+- [x] Remove Babel
+- [x] Switch to yarn
+- [x] Remove t411 tracker
+- [x] Remove Cpasbien tracker
+- [x] Add 1337x tracker
+- [x] Add Torrent9 tracker
+- [x] Cleanup
+
 ### v1.0.3
 
 - [x] Update README & CHANGELOG.